fix nextjs-config tests
