+++ packages/nextjs-config/package.json
@@ -21,6 +21,7 @@
     "lint": "eslint src",
     "lint:fix": "eslint src --fix",
     "build": "rslib build",
+    "test:unit": "jest",
     "dev": "rslib build --watch",
     "package": "pnpm pack --out $PACKAGE_DEST/%s.tgz"
   },
@@ -48,11 +49,12 @@
     "nextjs"
   ],
   "devDependencies": {
+    "@posthog-tooling/rollup-utils": "workspace:*",
+    "@posthog-tooling/tsconfig-base": "workspace:*",
+    "@rslib/core": "catalog:",
     "@types/node": "^22.15.23",
     "next": "^12.1.0",
-    "@posthog-tooling/tsconfig-base": "workspace:*",
-    "@posthog-tooling/rollup-utils": "workspace:*",
-    "@rslib/core": "catalog:"
+    "ts-jest": "catalog:"
   },
   "peerDependencies": {
     "next": ">12.1.0"
