+++ packages/nextjs-config/src/utils.ts
@@ -1,5 +1,6 @@
 import path from 'path'
 import fs from 'fs'
+import { spawn } from 'child_process'
 
 export function resolveBinaryPath(envPath: string, cwd: string, binName: string): string {
   const envLocations = envPath.split(path.delimiter)
@@ -38,4 +39,37 @@ const getLocalPaths = (startPath: string): string[] => {
   }
 
   return paths
+}
+
+export async function callPosthogCli(args: string[], env: NodeJS.ProcessEnv, verbose: boolean): Promise<void> {
+  let binaryLocation
+  try {
+    binaryLocation = resolveBinaryPath(process.env.PATH ?? '', __dirname, 'posthog-cli')
+  } catch (e) {
+    throw new Error(`Binary ${e} not found. Make sure postinstall script has been allowed for @posthog/cli`)
+  }
+
+  if (verbose) {
+    console.log('running posthog-cli from ', binaryLocation)
+  }
+
+  const child = spawn(binaryLocation, [...args], {
+    stdio: verbose ? 'inherit' : 'ignore',
+    env,
+    cwd: process.cwd(),
+  })
+
+  await new Promise<void>((resolve, reject) => {
+    child.on('close', (code) => {
+      if (code === 0) {
+        resolve()
+      } else {
+        reject(new Error(`Command failed with code ${code}`))
+      }
+    })
+
+    child.on('error', (error) => {
+      reject(error)
+    })
+  })
 }
