+++ pnpm-workspace.yaml
@@ -1,17 +1,16 @@
 packages:
-  # all packages in direct subdirs of packages/
-  - "packages/*"
-  - "tooling/*"
+  - packages/*
+  - tooling/*
 
 catalog:
-  tsup: "^8.5.0"
-  typescript: "^5.5.4"
-  rollup: "^4.44.1"
-  tslib: "^2.5.0"
-  jest: "^29.7.0"
-  jest-environment-jsdom: "^29.7.0"
-  jest-environment-node: "^29.7.0"
-  jest-expo: "^47.0.1"
-  "@types/jest": "^29.7.0"
-  ts-jest: "^29.0.0"
-  "@rslib/core": "^0.10.5"
+  tsup: ^8.5.0
+  typescript: ^5.5.4
+  rollup: ^4.44.1
+  tslib: ^2.5.0
+  jest: ^29.7.0
+  jest-environment-jsdom: ^29.7.0
+  jest-environment-node: ^29.7.0
+  jest-expo: ^47.0.1
+  '@types/jest': ^29.7.0
+  ts-jest: 29.4.0
+  '@rslib/core': ^0.10.5
